# api-spec/openapi/swagger/delegatee/v1/service.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": "version not set"
  },
  "paths": {
    "/v1/artifact": {
      "post": {
        "tags": [
          "DelegateeService"
        ],
        "description": "RegisterArtifact stores a compiler artifact templates may reference. Idempotent on its id.",
        "operationId": "DelegateeService_RegisterArtifact",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterArtifactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "a successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterArtifactResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/v1/artifact/{id}": {
      "get": {
        "tags": [
          "DelegateeService"
        ],
        "operationId": "DelegateeService_GetArtifact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetArtifactResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/v1/delegate": {
      "post": {
        "tags": [
          "DelegateeService"
        ],
        "description": "RegisterDelegation registers a watch: the daemon acts on coins sent to\nthe address the template derives for variables, until expires_at. The\ntemplate must have exactly one input; the wallet compares the address\nwith its own. Registering an active watch again returns it unchanged.",
        "operationId": "DelegateeService_RegisterDelegation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDelegationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "a successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDelegationResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/v1/delegate/{address}": {
      "get": {
        "tags": [
          "DelegateeService"
        ],
        "description": "GetDelegation returns the watch of an address (an active one first, else\nthe newest) with its spendable VTXOs (live from the arkd indexer) and its\nrenewal history.",
        "operationId": "DelegateeService_GetDelegation",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDelegationResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/v1/info": {
      "get": {
        "tags": [
          "DelegateeService"
        ],
        "description": "GetInfo returns the keys a wallet instantiates a template with.",
        "operationId": "DelegateeService_GetInfo",
        "responses": {
          "200": {
            "description": "a successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetInfoResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/v1/template": {
      "get": {
        "tags": [
          "DelegateeService"
        ],
        "description": "ListTemplates lists the active templates, without their documents.",
        "operationId": "DelegateeService_ListTemplates",
        "responses": {
          "200": {
            "description": "a successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTemplatesResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DelegateeService"
        ],
        "description": "RegisterTemplate stores a template the engine accepts. It is active at once. Idempotent on its id.",
        "operationId": "DelegateeService_RegisterTemplate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "a successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterTemplateResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/v1/template/{id}": {
      "get": {
        "tags": [
          "DelegateeService"
        ],
        "operationId": "DelegateeService_GetTemplate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTemplateResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Any": {
        "type": "object",
        "description": "Any contains an arbitrary schema along with a URL to help identify the type of the schema.",
        "properties": {
          "@type": {
            "type": "string",
            "description": "A URL/resource name that uniquely identifies the type of the schema."
          }
        }
      },
      "Artifact": {
        "title": "Artifact",
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "integer",
            "format": "int64"
          },
          "document": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        }
      },
      "Asset": {
        "title": "Asset",
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "uint64"
          },
          "assetId": {
            "type": "string"
          }
        }
      },
      "Delegation": {
        "title": "Delegation",
        "type": "object",
        "description": "Delegation is an instance of a template.",
        "properties": {
          "address": {
            "type": "string",
            "description": "address is slot 0's address for a watch, empty otherwise."
          },
          "createdAt": {
            "type": "integer",
            "format": "int64"
          },
          "expiresAt": {
            "type": "integer",
            "description": "expires_at is 0 when the delegation never expires.",
            "format": "int64"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "parentId": {
            "type": "integer",
            "description": "parent_id is the delegation whose settled transaction advertised this one, 0 for a watch.",
            "format": "int64"
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Slot"
            }
          },
          "status": {
            "type": "string",
            "description": "status is active, cancelled, expired or done."
          },
          "templateId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64"
          },
          "variables": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "GetArtifactRequest": {
        "title": "GetArtifactRequest",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "GetArtifactResponse": {
        "title": "GetArtifactResponse",
        "type": "object",
        "properties": {
          "artifact": {
            "$ref": "#/components/schemas/Artifact"
          }
        }
      },
      "GetDelegationRequest": {
        "title": "GetDelegationRequest",
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          }
        }
      },
      "GetDelegationResponse": {
        "title": "GetDelegationResponse",
        "type": "object",
        "properties": {
          "delegation": {
            "$ref": "#/components/schemas/Delegation"
          },
          "renewals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Renewal"
            }
          },
          "vtxos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vtxo"
            }
          }
        }
      },
      "GetInfoRequest": {
        "title": "GetInfoRequest",
        "type": "object"
      },
      "GetInfoResponse": {
        "title": "GetInfoResponse",
        "type": "object",
        "properties": {
          "delegatePubkey": {
            "type": "string"
          },
          "emulatorPubkey": {
            "type": "string"
          },
          "encryptionPubkey": {
            "type": "string",
            "description": "encryption_pubkey seals template secrets for this daemon; empty without an encryption key."
          },
          "network": {
            "type": "string"
          },
          "serverPubkey": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "GetTemplateRequest": {
        "title": "GetTemplateRequest",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "GetTemplateResponse": {
        "title": "GetTemplateResponse",
        "type": "object",
        "properties": {
          "template": {
            "$ref": "#/components/schemas/Template"
          }
        }
      },
      "ListTemplatesRequest": {
        "title": "ListTemplatesRequest",
        "type": "object",
        "description": "ListTemplatesRequest has no fields: the public list is always the active one."
      },
      "ListTemplatesResponse": {
        "title": "ListTemplatesResponse",
        "type": "object",
        "properties": {
          "templates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Template"
            }
          }
        }
      },
      "Param": {
        "title": "Param",
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "RegisterArtifactRequest": {
        "title": "RegisterArtifactRequest",
        "type": "object",
        "properties": {
          "document": {
            "type": "string"
          }
        }
      },
      "RegisterArtifactResponse": {
        "title": "RegisterArtifactResponse",
        "type": "object",
        "properties": {
          "artifact": {
            "$ref": "#/components/schemas/Artifact"
          }
        }
      },
      "RegisterDelegationRequest": {
        "title": "RegisterDelegationRequest",
        "type": "object",
        "properties": {
          "expiresAt": {
            "type": "integer",
            "description": "expires_at is a unix time after which the watch stops; 0 never expires.",
            "format": "int64"
          },
          "templateId": {
            "type": "string"
          },
          "variables": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "RegisterDelegationResponse": {
        "title": "RegisterDelegationResponse",
        "type": "object",
        "properties": {
          "delegation": {
            "$ref": "#/components/schemas/Delegation"
          }
        }
      },
      "RegisterTemplateRequest": {
        "title": "RegisterTemplateRequest",
        "type": "object",
        "properties": {
          "document": {
            "type": "string"
          }
        }
      },
      "RegisterTemplateResponse": {
        "title": "RegisterTemplateResponse",
        "type": "object",
        "properties": {
          "template": {
            "$ref": "#/components/schemas/Template"
          }
        }
      },
      "Renewal": {
        "title": "Renewal",
        "type": "object",
        "properties": {
          "attemptedAt": {
            "type": "integer",
            "format": "int64"
          },
          "commitmentTxid": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "outpoints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "Slot": {
        "title": "Slot",
        "type": "object",
        "description": "Slot is one template input of a delegation.",
        "properties": {
          "name": {
            "type": "string"
          },
          "onchain": {
            "type": "boolean"
          },
          "outpoint": {
            "type": "string",
            "description": "outpoint is \"txid:vout\" when bound, empty for a watch."
          },
          "tapscripts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Status": {
        "title": "Status",
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Any"
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Template": {
        "title": "Template",
        "type": "object",
        "properties": {
          "artifactIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "integer",
            "format": "int64"
          },
          "document": {
            "type": "string",
            "description": "empty in list responses"
          },
          "id": {
            "type": "string"
          },
          "params": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Param"
            }
          },
          "status": {
            "type": "string"
          }
        }
      },
      "Vtxo": {
        "title": "Vtxo",
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "uint64"
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "createdAt": {
            "type": "integer",
            "format": "int64"
          },
          "expiresAt": {
            "type": "integer",
            "format": "int64"
          },
          "onchain": {
            "type": "boolean"
          },
          "outpoint": {
            "type": "string"
          },
          "preconfirmed": {
            "type": "boolean"
          },
          "renewableAt": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "DelegateeService"
    }
  ]
}