# internal/infrastructure/db/postgres/sqlc/queries/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package queries

import (
	"database/sql"
	"encoding/json"
	"time"
)

type Artifact struct {
	ID        string
	Document  []byte
	CreatedAt time.Time
}

type Delegation struct {
	ID             int64
	Fingerprint    string
	Address        string
	TemplateID     string
	Variables      json.RawMessage
	ParentID       sql.NullInt64
	ExpiresAt      sql.NullTime
	DelegatePubkey string
	Slots          json.RawMessage
	Status         string
	CreatedAt      time.Time
	UpdatedAt      time.Time
}

type Renewal struct {
	ID             int64
	DelegationID   int64
	Outpoints      []string
	CommitmentTxid string
	Success        bool
	Error          string
	AttemptedAt    time.Time
}

type Settlement struct {
	Txid         string
	SpentBy      string
	DelegationID int64
	Tx           []byte
	Outpoints    json.RawMessage
	Sources      [][]byte
	Coins        []string
	Checkpoints  []string
	Finals       []string
	Landed       bool
	CreatedAt    time.Time
}

type Template struct {
	ID          string
	Document    []byte
	ArtifactIds []string
	Params      json.RawMessage
	Status      string
	Failures    int32
	Trusted     bool
	CreatedAt   time.Time
	UpdatedAt   time.Time
}